# Improve RoomList performance via side-stepping React