--- a/src/components/structures/TimelinePanel.js
+++ b/src/components/structures/TimelinePanel.js
@@ -523,9 +523,7 @@ var TimelinePanel = React.createClass({
                 this.props.timelineSet.room.setUnreadNotificationCount('highlight', 0);
                 dis.dispatch({
                     action: 'on_room_read',
-                    payload: {
-                        room: this.props.timelineSet.room
-                    }
+                    room: this.props.timelineSet.room,
                 });
             }
         }