Previously we regenerated the whole RoomList state on pretty much any event from the js-sdk, causing React to them re-render the whole RoomList.  This is incredibly heavy (~350ms on my account), and chews a lot of CPU and makes the UI laggy.

This switches the common paths of handling js-sdk events for selecting rooms, `Room.timeline`, `Room.receipt` and a few easy ones like `Room.name`, `RoomMember.name`, `RoomState.members` to side-step React and prod the specific relevant RoomTile or RoomSubList to update themselves rather than have to poke them via React props.  Rarer ones like `Room.tags`, `Room`, `deleteRoom`, `m.direct` still cause a full regeneration of the room list.

It does this by dispatching an event to them using a new ConstantTimeDispatcher (CTD), which the tiles & sublists use to register themselves to be directly prodded for these events.  CTD works by registering listeners for both an event type and an argument, meaning that only the right listener gets woken up for a given argument rather than iterating over all of them (which could be thousands of RoomTiles).

In a test jig benchmark (changing the name of a single roomtile in a list of 2000), it took 350ms (180ms in prod webpack) of processing when updating the tile by generating a new list, versus 3.5ms when doing a constant time dispatch on the right tile.

Alternative approaches could be to:
 * have RoomTiles subscribe directly to the events emitted from Room objects in the js-sdk (rather than the events as re-emitted from MatrixClient).  @dbkr suggested this after the ConstantTimeDispatcher stuff was already in existence, and given CTD feels like a vaguely useful tool to have around, I haven't had the heart to yank it out again, although if this works it'd simplify the dispatching a bit.
 * have RoomTiles register themselves as references on their RoomSubList, so parents can prod them directly to re-render when needed.
 * use immutable state (e.g. Immutable.js) to force the right React components to re-render on demand, at the expense of generating new data every time anything changes (meaning shouldComponentUpdate can just do a trivial shallow comparison).  This is presumably the most 'reacty' solution.
 
Twinned with https://github.com/vector-im/riot-web/pull/3654 (although oh my god we need to move RoomSubList* into react-sdk)