@ara4n 2017-04-18
not entirely sure why travis is failing; tests pass locally.  travis' logs say:

```
[1A[2KPhantomJS 2.1.1 (Linux 0.0.0) joining a room over federation should not get stuck at a spinner FAILED
	Expected null to exist
	assert@/home/travis/build/matrix-org/matrix-react-sdk/riot-web/test/all-tests.js:4946:28
	toExist@/home/travis/build/matrix-org/matrix-react-sdk/riot-web/test/all-tests.js:2309:29
	/home/travis/build/matrix-org/matrix-react-sdk/riot-web/test/all-tests.js:9371:53
```

...which isn't very helpful.