@Kegsay 2017-04-19
I've just read your summary for now, not the code. Broadly speaking what you're doing makes a lot of sense, though I would've preferred option 2:

> have RoomTiles register themselves as references on their RoomSubList, so parents can prod them directly to re-render when needed

Option 1 will end up being slow as the number of rooms increases as it still needs to loop every single `RoomTile` when any event fires, which is not free.

Option 3 as you say, generates a lot of junk, and I'm unconvinced that the GC churn this will create won't become a serious problem if we did this everywhere.

I'll update this comment when I've had a look at the code.