@ara4n 2017-04-19
After discussing IRL: kegan & matthew quite liked the idea of using refs, as they keep a parent->child relationship; dave quite liked using dispatched events (but would prefer not having CTD in there).  Conclusion is that nobody cares enough to necessarily require the PR to be rewritten, so if it works here, keep it as is for now, but file a maintenance bug for restructuring it to perhaps use refs or a non-CTD dispatcher in future.